per-platform packages
