+++ .github/workflows/get_artifact_dir_name.js
@@ -1,7 +1,7 @@
 import * as fs from "node:fs";
 import * as os from "node:os";
 
-import { platformName } from "#cli/bins";
+import { platformName } from "#dev/paths";
 
 // Pass artifactDirName to subsequent GitHub actions
 fs.appendFileSync(
