+++ packages/@rescript/darwin-arm64/bin/bsc.exe
@@ -1,0 +1,1 @@
+../../../../darwinarm64/bsc.exe
