+++ packages/@rescript/darwin-arm64/bin/ninja.exe
@@ -1,0 +1,1 @@
+../../../../darwinarm64/ninja.exe
