+++ packages/@rescript/darwin-arm64/bin/rescript-editor-analysis.exe
@@ -1,0 +1,1 @@
+../../../../darwinarm64/rescript-editor-analysis.exe
