+++ packages/@rescript/darwin-arm64/bin/rescript.exe
@@ -1,0 +1,1 @@
+../../../../darwinarm64/rescript.exe
