+++ packages/@rescript/darwin-x64/bin/bsc.exe
@@ -1,0 +1,1 @@
+../../../../darwin/bsc.exe
