+++ packages/@rescript/darwin-x64/bin/ninja.exe
@@ -1,0 +1,1 @@
+../../../../darwin/ninja.exe
