+++ packages/@rescript/darwin-x64/bin/rescript-editor-analysis.exe
@@ -1,0 +1,1 @@
+../../../../darwin/rescript-editor-analysis.exe
