+++ packages/@rescript/darwin-x64/bin/rescript-tools.exe
@@ -1,0 +1,1 @@
+../../../../darwin/rescript-tools.exe
