+++ packages/@rescript/darwin-x64/bin/rescript.exe
@@ -1,0 +1,1 @@
+../../../../darwin/rescript.exe
