+++ packages/@rescript/darwin-x64/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@rescript/linux-arm64",
+  "name": "@rescript/darwin-x64",
   "version": "12.0.0-alpha.12",
-  "description": "ReScript binaries for Linux ARM64",
+  "description": "ReScript binaries for MacOS x86_64",
   "type": "module",
   "repository": {
     "type": "git",
@@ -14,10 +14,10 @@
     "node": ">=20.11.0"
   },
   "os": [
-    "linux"
+    "darwin"
   ],
   "cpu": [
-    "arm64"
+    "x64"
   ],
   "files": [
     "bin"
