+++ packages/@rescript/linux-arm64/bin/bsc.exe
@@ -1,0 +1,1 @@
+../../../../linuxarm64/bsc.exe
