+++ packages/@rescript/linux-arm64/bin/ninja.exe
@@ -1,0 +1,1 @@
+../../../../linuxarm64/ninja.exe
