+++ packages/@rescript/linux-arm64/bin/rescript-editor-analysis.exe
@@ -1,0 +1,1 @@
+../../../../linuxarm64/rescript-editor-analysis.exe
