+++ packages/@rescript/linux-arm64/bin/rescript-tools.exe
@@ -1,0 +1,1 @@
+../../../../linuxarm64/rescript-tools.exe
