+++ packages/@rescript/linux-x64/bin/bsc.exe
@@ -1,0 +1,1 @@
+../../../../linux/bsc.exe
