+++ packages/@rescript/linux-x64/bin/ninja.exe
@@ -1,0 +1,1 @@
+../../../../linux/ninja.exe
