+++ packages/@rescript/linux-x64/bin/rescript-editor-analysis.exe
@@ -1,0 +1,1 @@
+../../../../linux/rescript-editor-analysis.exe
