+++ packages/@rescript/linux-x64/bin/rescript-tools.exe
@@ -1,0 +1,1 @@
+../../../../linux/rescript-tools.exe
