+++ packages/@rescript/linux-x64/bin/rescript.exe
@@ -1,0 +1,1 @@
+../../../../linux/rescript.exe
