+++ packages/@rescript/win32-x64/bin/bsc.exe
@@ -1,0 +1,1 @@
+../../../../win32/bsc.exe
