+++ packages/@rescript/win32-x64/bin/ninja.exe
@@ -1,0 +1,1 @@
+../../../../win32/ninja.exe
