+++ packages/@rescript/win32-x64/bin/rescript-editor-analysis.exe
@@ -1,0 +1,1 @@
+../../../../win32/rescript-editor-analysis.exe
