+++ packages/@rescript/win32-x64/bin/rescript-tools.exe
@@ -1,0 +1,1 @@
+../../../../win32/rescript-tools.exe
