+++ packages/@rescript/win32-x64/bin/rescript.exe
@@ -1,0 +1,1 @@
+../../../../win32/rescript.exe
