+++ packages/@rescript/win32-x64/bin/rewatch.exe
@@ -1,0 +1,1 @@
+../../../../win32/rewatch.exe
