+++ scripts/copyExes.js
@@ -7,7 +7,7 @@
 import * as child_process from "node:child_process";
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { platformDir } from "#cli/bins";
+import { platformDir } from "#dev/paths";
 import { compilerBinDir, ninjaDir, rewatchDir } from "#dev/paths";
 
 fs.mkdirSync(platformDir, { recursive: true });
